Begin a project/tag listing as well
It is useful to have a project/tag listing option. This starts such a listing. I
do a bit of refactoring to move some formatting and such around. Plus I start to
move the project and tag specific things into their own files. The listing is
created correctly (I think), but the actions do not yet work. That will need
some work.