Add actions to start and stop a task as in #1
It is possible to start and stop a specific task this way. This should may be
extended to allow several tasks, though I'm not sure of how useful that is.
Also, this suggests that we display which tasks are started. A nice syntax file
would be a good way to do this.