Skeleton of how to annotate using external tasks
As suggested in #1 it would be very useful to annotate a task with the results
of another unite source. To do this we can add an annotate action to the common
kind which is present in nearly all sources. I need to do something to allow for
selecting the task(s) to use by opening another unite window that can only
select tasks.