Add annotate action to all sources for #1
Being able to add an annotation to any task is a very useful feature. I've added
this functionality. It allows the selection of any number of items and adds them
to any number of tasks as an annotation.

To do this I added a new action to the common kind which most kinds inherit
from. This then activates a taskwarrior source to select tasks to annotate.
There are some limitations to this approach. For example, we don't use the same
source for normal task listing and for task selecting leading to an issue with
adding matchers and sorters and such. The matchers and sorters have to be added
to both or we end up not being able to filter as expected when selecting a task
to annotate. I also limit the action available when selecting to only annotate.
I think this is a good limitation however.