Add an option to infer the project
I often have directory names that correspond to the project I'm working on. It
is useful to only see the tasks for the current project so I've added an infer
option. Also, I've prototyped out a bit of code to load data from a
.unite-taskwarrior file. If that file exists it will be used to get the current
project. This option will only show tasks if their projects match the inferred
project name.