Add description to note file as in #10
This will create a note file with the description as the first line. It might be
better to add an empty line afterwards, but that is kinda up to the user as to
how they want stuff formatted. Another useful option might be to have a function
which generates the function so it is possible to create markdown headers, vim
wiki style entries, or whatever. However, this works as a first pass.