Fix #9
I changed the interface to filter. This should fix up the issue with projects
not being listed. However, this is pretty basic at the moment and I'm not sure
it can do things like list all tags for a project 'Unite
taskwarrior/tags:$project' correctly. However, this allows the basic
functionality which is still useful.