Fix #6
There was a bug with creating tasks inside vim where it would add tasks that
started with ( ) and would not respect tags and projects. This was due to
quoting and bad arguments to the command line. This is now fixed by flattening
and expanding lists as well as removing the empty filter string to when calling
taskwarrior.