Correct when paths are expanded for #10
We should always expand the path when we need to create the directory for note
files. This should happen during init as we won't be building files before that.
This can fail if we change the directory or attempt to make notes before running
init, but this is not standard use case. In fact you'd have to write code to do
that and it is not supported.