Add optional keybindings
It will be useful to new users to use standardish key bindings in the unite
interface. Some of the bindings are likely to interfere with useful standard
unite ones, but many do not. So we will try to define such useful bindings.
Right now there are only a few <TAB> for toggle, d for do, D for delete and <CR>
for view, and P for edit project. This list may expand or shrink and I many
increase the complexity of the configuration later. Right now as these are still
very experimental and I'm not sure if I like them I have them as optional and
defaulting to off. They are activated by setting the
g:unite_taskwarrior_define_mappings variable to 1, which defaults to 0 for now.