Add a source for new tasks
This way it is possible to open and manipulate a new task while in the standard
task listing. This is same sort of idea as the file_new source in unite. It
would also be a good idea to allow stuff like having the project source open a
taskwarrior/new source where the project is set to the desired one by default.
This is tricky but would be useful.