Staring work on multi line annotations
Multi line annotations will be a very useful possible replacement for note
files. This is not yet available in taskwarrior binaries but is coming up soon
and this is preparing for it. The basic idea for this is to create a temporary
buffer, possibly not attached to a file, for writing. Once the buffer is quit
the selected tasks are annotated. This would be both an action in the unite
viewer as well as a command that could be called.