Correct bugs in modifying tags
Before this would modify the description and not the tags, which is clearly
wrong and bad. This is now fixed however, it does rely upon the tag abbreviation
being + and has a limit of up to 20 modifications do to the number of arguments
that can be given to vim functions. It works though.