Let annotate use a scratch buffer for #16
For #16 we will need to open up a temp buffer that we can write to. This will
then be read on close and stored as annotation. Since the current taskwarrior
does not support multiline annotations we convert all tabs and newlines into
spaces. This should allow for use with current taskwarrior, but put is in a good
spot for the future versions.