Show annotations and alter configuration
Configuration was taking up lots of lines as well as creating tons of global
variables. This removes such things by adding a function to get configuration
information. While you can still use global variables it is possible to use
function calls to configure. In addition, we now only access the function for
most configuration work.

In addition, we now show annotations for a task along with the task. The format
is not ideal and it is only done when showing with the standard formatter, but
it is a start. This also makes the source a multiline source in order to work.