Add a source and actions for dependencies
It is useful to be able to add dependencies at times. This is the first step
along the way to fully display and edit them. This adds a source like the
annotate one in that it allows a task to be modified with dependencies. This
does imply that I can simplify the code a bit and make the two things use the
same code. I find creating dependencies to be a key part of organizing my
workflow and how I will spend my time.