Update to handle JSON array output
As noted in #17 this cannot handle taskwarrior exporting proper a JSON array of
tasks. This will become standard in newer versions of taskwarrior. This fixes
that issue by attempting to detect if the first character is a '[', which means
JSON array and then parsing the entire output. If not it uses the old behavior
of parsing each line individually. It may be better to try to parse as JSON and
then fail if it is not and then fall back to the previous method.