Initial attempt at adding a history action
It is useful in several cases to be able to open the previous unite buffer. This
was suggested in #14. This is very useful in navigating between the buffers. The
current problem with this is that I can't bind it to <BS> from within the plugin
but I can do it from within my own vimrc, which is odd. To use this add
something like:

nnoremap <silent><buffer><expr> <BS>  unite#do_action('previous')

to ftplugin/unite.vim.