Attempt to correct generation of new tasks
With json.array = on the export of a single task always produces a list. This
causes an issue when creating a new task as it assume that the output is just a
single json dictionary and not a list. This fixes newest so it only returns the
single newest task.