Add ability to pass custom filter to taskwarrior
This allows the usage of a filter which will override the default filter. This
is something that I have wanted for a bit and will prove useful. The
modification is to place the addition of the standard filter inside the call to
build a filter. If we are passed a custom filter we use that instead.