Rework call function interface
This function should take either a string or an array like the underlaying
vimproc#system call. This is because in some cases it will be hard to split up
the arguments to a system call. For example, consider the case of doing an add
like: 'task add Something "tag:a with space"' or something that needs to be
quoted. I'd rather leave that parsing to something else instead of just
splitting on whitespace and hoping for the best. In addition, this makes it
easier to generate some calls.