Partially working context source
This can now list and count contexts correctly. We can get a list of all defined
contexts as well as the counts of the non completed tasks. I may also want to
add a filter for not deleted as well. This doesn't add any of the features
related to setting to none and there are no tests yet.