Add a similar tasks action
This will list all tasks which are similar to the selected tasks. We define
similar to be all things in the same project with at least one tag in common. I
feel like this is a reasonable selection of tasks. It may select too much in
that this does not respect the default filter but maybe it should.