Attempt to list only tasks for given context
This should list only the tasks for the given context once it is used. This
works by building a filter for export to use. This is because export does not
respect the context filter once set. I don't know if this is a bug in
taskwarrior but it is surprising. This attempts to build a filter that respects
the context and the default filter. This also modifies the taskwarrior source
and taskwarrior select to be able to use a single string filter more correctly.