Add a function to load notes files
This will load up all notes files and create a simple but useful dictionary
about them. Right now the formatting is basically a fixed part of the dictionary,
which is lazy and may not be a good idea. However, this does provide a simple
way to load the known notes files.