Add a notes source as per #22
This provides a simple source to display the notes files. This allow a user to
see the notes and open them as needed. The source is a bit limited at the moment
as we cannot view the associated task, which would be helpful, or delete the
file which is dangerous but useful. However, this is a good start to #22. In
addition, this highlights an issue where 'old' notes files are not deleted or
archived.

Right now my thoughts are to deal with this outside of unite-taskwarrior. As I
can and do complete tasks outside of vim the only reliable way to deal with this
is to use a hook built in to taskwarrior.