Add a function to get current context
This will be used when adding a flag that forces the task list to respect the
current context. It is similar to loading all tasks, but it does not get counts
and only gets the current context. It may be able to speed it up by not using
the _show command but instead the context command.