Add test and fix bugs in configuration system
This test out the newer configuration system. In doing so I found some bugs
about setting and storing values. Also, it is useful for testing and maybe in
practice to be able to rest all configuration values. I've added a function for
that. It's also nice in that we can now recover the defaults at any time.