Add test for calling taskwarrior
This adds a few very simple tests to make sure we can call taskwarrior both
directly and using a task object. In doing so I found a bug with using nested
lists. This is now fixed. I worry that some of these test may be fragile with
respect to number, but hopefully reseting the task data will keep the resulting
ids the same.