Test for parsing and building
This tests parsing and building of task data structures. Here we try to parse a
variety of inputs. I noticed that there was logic in select which belongs in
parse, so I copied it over there. I still need to clean up select so it does not
duplicate any logic.