More building tests
This adds some more building tests and make the building procedure better able
to deal with missing dependencies. These are not always around and so we should
be able to work with that. This does produce an empty description in that case.
It may be best to prevent actions that involve such a task as it is pretty
meaningless.