Another attempt at fixing testing
I found some issues where the newest vimproc didn't work in testing. It does
seem to work when I call the functions by hand so I have no idea what's up
there. Thus we use the newest vimproc version that vim-flavor can parse, which
is 2.*, while the newest is ~8.0. Not ideal. Also, I removed the compile step
in the rakefile and placed it into travis configuration, this way the compiling
can be done correctly on travis, but we don't have to worry about platform
detection in the rakefile.