Maybe finally fix the testing issue
Ok so it seems that vimflavor uses the files that it places into the current
directory under .vim-flavor for testing. However, it does not do this until we
run test. Normally this is ok, but vimproc needs to be compiled. If we compile
it prior to running test the compiled binary still doesn't get copied over, for
unclear reasons. So to fix this we run a dummy test that does nothing, but does
cause the files to be copied and then compile vimproc as part of the setup. We
can then run tests safely. I hope anyway.