Correct issues with selection tests
This were not getting reset because rake was never running. Though I struggle to
see how rake was not running when rake ci is the test command but ok, whatever.
Somehow rake wasn't available while a test was running. I previously added it to
the Gemfile which fixed that issue. Now I correct the test which was meant to
fail in the hopes that now everything will actually pass.