Add formatters for tasks
This adds a better taskwiki and simple formatter. This was suggested in #36 as
well as #37 as improvements to the display. The first will allow for better
integration of tasks into taskwiki and the second will allow for a more standard
tag display. Also, I have made the simple format always 80 wide. This is nice
so the tags are always right aligned.

The formatters are different than previous things, notably these take options
about the sizes of things to control how they will display stuff. So we can
assign less than 10 (or whatever the default is) to projects to allow for a more
full display. Also, by limiting to 80 characters the simple one should always be
displayable, no truncation by unite needed. The taskwiki format does not have
this, though I don't know if it should.