Add some code for integrating with taskwiki
This adds some of the code needed for #36. We need to be able to find and open
taskwiki files for a given task. This involves checking the annotations, then
the project and then finally using a default one. I've added a bit of code to do
that as well as a couple tests for it.