Add showing of annotations back
This will now show annotations in the task view if show_annotations is on and we
are using the simple view. Simple isn't really a good name of it anymore, maybe
standard would be better. This indents using 2 characters, which is the default
for taskwarrior, but ideally this could respect the indentation rules configured
for taskwarrior.