Move to use standard formatters
This moves code around so we place the formatters into the formatter module and
use them for tags, contexts and projects. This makes them standard as well as
helps limit the size of the produced views. I like the improved consistency as
well as having a nice standard formatters for everything. This should make the
displays nicer overall I think.