Indicate active status with *
Since showing active tasks is really useful, at least for me, we add this
feature. The simple formatting already shows a leading '-' for standard tasks,
and a '+' for a task with notes. We can then move to use '*' for active tasks. I
like this feature as I find it really helpful.