Create some basic filter logic
This adds the basic filtering abilities. A filter is something that can take a
series of projects, tags, raw filters and contexts and build string that is
useful to taskwarrior. This will hopefully centralize the generation of filters.
It seems most attempts at building them need the same sort of logic so this is a
good idea, I think.