Use new filter system for context filters
So speed up switching we don't use the context method of the filter. This
however poses an issue when respect context is set. The filter will not match
all tasks in the context because the current context filter will be added as
well the one for the selected filter. This can be overcome by ignore the filter
when switching.