Add tests for settings manipulation
We can now get settings as needed for #32. This is a set of some basic test for
the functions of setting and getting taskwarrior configuration values. There is
an odd asymmetry in this where we can't use rc.name to set the value, but we
must use it to get. I don't get why, but that is just how it is.