Add simple config source for #32
This adds a very simple configuration source and kind for this module. You can
interactively find and edit configuration values. This is pretty bare bones at
the moment but there are some possible extensions, for example showing which are
default and including a description of any value would be useful. Or
alternatively the option to open help about item.