Improve annotation formatting for #35
The current annotation formatting is not very good. This attempts to improve it
by using the simple format. We show the annotation as a single line then show a
modified simple formatted task below it. This should improve the annotation
formatting. Hopefully at least. This may also make it harder to read stuff as it
may be too much to read at once.