Add a function for previewing results
This is generally useful. For example when previewing the results of information
or when running a command we can open a preview window to view the results. This
allows us to integrate with AnsiEsc so we get colors and the like. The downside
is that this is a different behavior than before. Also, the previous one would
autoclose which was nice. This does not autoclose when moving between rows.