Add a try/catch to getting newest
Sometimes when getting the newest, this fails. The cause is that if forcecolor
is on this will create output which cannot be parsed by the current system. I
should do a couple things to fix this, notably improving the parsing. For now I
will simply stop the error from showing up. It does create the task, but should
not show this error.