Add some configuration variables for sorting
We will need a dictionary mapping name to sorting ordering, as well as a default
ordering. The default ordering is a dummy now I will probably change it back to
urgency later. We integrate our configuration with taskwiki stuff. That is we
can use the orderings defined there as well. We will also merge them with the
orderings defined for this to produce the overall one.