Add tests for parsing
We can now parse a sorting order string into a list of attributes. They will
contain the information needed to sort tasks by the specified order. It
basically splits up the parts by , and then stores the name and then the + or -
at the end as the ordering. It is fairly strict in requiring a + or -.