Allow sorting by date for #28
This now adds date sorting to the task sorter. This requires that both
attributes look like dates otherwise it will just use normal comparison. This
sorting is likely to be much slower than normal sorting but it does try to cache
the results. Hopefully it's not too bad. That is kinda the drawback to using
vimscript for everything.