Add a completion function for next status for #41
When toggling the next status we need to use a completion function to switch
between statuses. This adds a function which allows users to hit tab and have
the status on the command line switch through the status mapping. It is pretty
simple now but should do the trick.

Oddly this exposes a potential bug in neovim. In vim this works fine switching
between completed and pending. However, in neovim it does not. It appears that
the suggested word must always be longer than the one being completed. This is
strange behavior and should probably be reported to that project.